+++ wrapper/static/page.js
@@ -27,7 +27,11 @@ module.exports = function (req, res, url) {
 			title = 'Character Creator';
 			attrs = {
 				data: process.env.SWF_URL + '/cc.swf', // data: 'cc.swf',
-				type: 'application/x-shockwave-flash', id: 'char_creator', width: '100%', height: '100%',
+				type: 'application/x-shockwave-flash', 
+				id: 'char_creator', 
+				width: '960', 
+				height: '600', 
+				style:'display:block;margin-left:auto;margin-right:auto;',
 			};
 			params = {
 				flashvars: {
@@ -38,6 +42,42 @@ module.exports = function (req, res, url) {
 				},
 				allowScriptAccess: 'always',
 				movie: process.env.SWF_URL + '/cc.swf', // 'http://localhost/cc.swf'
+			};
+			break;
+		}
+		
+		case "/cc_browser": {
+			title = "CC Browser";
+			attrs = {
+				data: process.env.SWF_URL + "/cc_browser.swf", // data: 'cc_browser.swf',
+				type: "application/x-shockwave-flash",
+				id: "char_creator",
+				width: '100%', 
+				height: '600', 
+				style:'display:block;margin-left:auto;margin-right:auto;',
+			};
+			params = {
+				flashvars: {
+					apiserver: "/",
+					storePath: process.env.STORE_URL + "/<store>",
+					clientThemePath: process.env.CLIENT_URL + "/<client_theme>",
+					original_asset_id: query["id"] || null,
+					themeId: "family",
+					ut: 60,
+					appCode: "go",
+					page: "",
+					siteId: "go",
+					m_mode: "school",
+					isLogin: "Y",
+                                        retut: 1,
+                                        goteam_draft_only: 1,
+					isEmbed: 1,
+					ctc: "go",
+					tlang: "en_US",
+					lid: 13,
+				},
+				allowScriptAccess: "always",
+				movie: process.env.SWF_URL + "/cc_browser.swf", // 'http://localhost/cc_browser.swf'
 			};
 			break;
 		}
@@ -75,6 +115,7 @@ module.exports = function (req, res, url) {
 					'autostart': 1, 'isWide': 1, 'clientThemePath': process.env.CLIENT_URL + '/<client_theme>',
 				},
 				allowScriptAccess: 'always',
+				allowFullScreen: 'true',
 			};
 			break;
 		}
@@ -84,7 +125,72 @@ module.exports = function (req, res, url) {
 	}
 	res.setHeader('Content-Type', 'text/html; charset=UTF-8');
 	Object.assign(params.flashvars, query);
-	res.end(`<script>document.title='${title}',flashvars=${JSON.stringify(params.flashvars)}</script><body style="margin:0px">${toObjectString(attrs, params)
+	res.end(`
+	<head>
+		<script>
+			document.title='${title}',flashvars=${JSON.stringify(params.flashvars)}
+		</script>
+		<script>
+			if(window.location.pathname == '/player') {
+				function hideHeader() {
+					document.getElementById("header").style.display = "none";
+				}
+			} else if(window.location.pathname == '/go_full') {
+				function hideHeader() {
+					document.getElementById("header").style.display = "none";
+				}
+			}
+		</script>
+		<link rel="stylesheet" type="text/css" href="/pages/css/modern-normalize.css">
+		<link rel="stylesheet" type="text/css" href="/pages/css/global.css">
+		<style>
+			body {
+				background: #eee;
+			}
+		</style>
+	</head>
+	
+	<header id="header">
+		<a href="/"><h1 style="margin:0"><img id="logo" src="/pages/img/list_logo.png" alt="Wrapper: Offline"/></h1>
+		<nav id="headbuttons">
+			<a class="button_small" onclick="document.getElementById('file').click()">UPLOAD A MOVIE</a>
+			<div class="dropdown_contain button_small">
+				<div class="dropdown_button">CREATE A CHARACTER</div>
+				<nav class="dropdown_menu">
+					<h2>Comedy World</h2>
+					<a href="/cc?themeId=family&bs=adam">Guy (Adam)</a>
+					<a href="/cc?themeId=family&bs=eve">Girl (Eve)</a>
+					<a href="/cc?themeId=family&bs=bob">Fat (Bob)</a>
+					<a href="/cc?themeId=family&bs=rocky">Buff (Rocky)</a>
+					<hr>
+					<h2>Anime</h2>
+					<a href="/cc?themeId=anime&bs=guy">Guy</a>
+					<a href="/cc?themeId=anime&bs=girl">Girl</a>
+					<a href="/cc?themeId=ninjaanime&bs=guy">Guy (Ninja)</a>
+					<a href="/cc?themeId=ninjaanime&bs=girl">Girl (Ninja)</a>
+					<hr>
+					<h2>Peepz</h2>
+					<a href="/cc?themeId=cc2&bs=default">Lil Peepz</a>
+					<a href="/cc?themeId=chibi&bs=default">Chibi Peepz</a>
+					<a href="/cc?themeId=ninja&bs=default">Chibi Ninjas</a>
+				</nav>
+			</div>
+			<div class="dropdown_contain button_small">
+				<div class="dropdown_button">BROWSE CHARACTERS</div>
+				<nav class="dropdown_menu">
+					<a href="/cc_browser?themeId=family">Comedy World</a>
+					<a href="/cc_browser?themeId=anime">Anime</a>
+					<a href="/cc_browser?themeId=ninjaanime">Ninja Anime</a>
+					<a href="/cc_browser?themeId=cc2">Lil' Peepz</a>
+					<a href="/cc_browser?themeId=chibi">Chibi Peepz</a>
+					<a href="/cc_browser?themeId=ninja">Chibi Ninjas</a>
+				</nav>
+			</div>
+			<a href="/go_full" class="button_big">MAKE A VIDEO</a>
+		</nav>
+	</header>
+	
+	<body style="margin:0px" onload="hideHeader()">${toObjectString(attrs, params)
 		}</body>${stuff.pages[url.pathname] || ''}`);
 	return true;
 }